api/requirements.txt
fastapi
uvicorn
mangum
pydantic
google-genai
pandas
openpyxl
qrcode[pil]
Pillow
requests
